md_12/HW_fasapi/pyproject.toml
[tool.poetry]
name = "hw-fasapi"
version = "0.1.0"
description = ""
authors = ["takeRednotBlue <maxymklym1996@gmail.com>"]
readme = "README.md"
packages = [{include = "hw_fasapi"}]

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.21"
alembic = "^1.12.0"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.0"
fastapi = "^0.103.2"
uvicorn = {extras = ["standart"], version = "^0.23.2"}
email-validator = "^2.0.0.post2"
pydantic-extra-types = "^2.1.0"
phonenumbers = "^8.13.22"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
fastapi-jwt-auth = "^0.5.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
libgravatar = "^1.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
